// frontend/app/models/kupljena.ts
export class kupljenaNekretnina {
    opis: string;
    adresa: string;
    tip: string;
    sprat: string;
    kvadratura: string;
    sobe: string;
    namestenost: string;
    izdavanje: string;
    galerija: string[];
    // cena: string;
    cena:number;
    vlasnik: string;
    id: string;
    promo:boolean;
    kupac:string;
}